refactor: rename Countdown component
