+++ components/Countdown/index.ts
@@ -1,0 +1,1 @@
+export { default as Countdown } from "./Countdown";
