+++ components/Hours/Hour.tsx
@@ -1,5 +1,5 @@
 import { IHourData } from "@mauriciorobayo/pyptron";
-import Countdown from "components/countdown";
+import { Countdown } from "components/Countdown";
 import { ALL_DAY, AMERICA_BOGOTA } from "lib/utils";
 import { useEffect, useState } from "react";
 import styled from "styled-components";
